abstract out not found content
